requirements.txt
accelerate
bitsandbytes
cpm_kernels
datasets
fire
flask
flask_cors
line-bot-sdk
googletrans==4.0.0-rc1
grpcio
loguru
opencc-python-reimplemented
pandas
peft
Pillow
pyngrok
sentencepiece
tqdm
transformers
wheel